pyproject.toml
[project]
name = "first-prompt-macher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "google-cloud-bigquery>=3.35.1",
    "google-genai>=1.28.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "tqdm>=4.67.1",
]
